pyproject.toml
[tool.black]
line-length = 88
target-version = ['py36']
include = '\.pyi?$'
exclude = '''
/(
    \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
  | data
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3

[tool.poetry]
name = "hack"
version = "0.2.3"
description = "Creating Hardware Component Knowledge Bases"
authors = ["Luke Hsiao <lwhsiao@stanford.edu>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.7"
Pillow = "^8.1.2"
fonduer = "^0.8.2"
matplotlib = "^3.3.4"
numpy = "^1.11.0,<1.20.0"
quantiphy = "^2.13.0"
scipy = "^1.6.1"
seaborn = "^0.11.1"
snorkel = "^0.9.5,<0.9.8"
statsmodels = "^0.12.2"
torch = "^1.3.0"
torchvision = "^0.9.0"
tqdm = "^4.59.0"

[tool.poetry.dev-dependencies]
black = "^20.8b1"
flake8 = "^3.9.0"
isort = "^5.8.0"
jupyter = "^1.0.0"
pre-commit = "^2.11.1"
pytest = "^6.2.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
